add a couple descriptor-related examples (#14)
One of the two added examples clearly fits into the "narrowing" category
(it uses a descriptor, but it's about narrowing behavior.) The other one
didn't seem to clearly fit in any existing category -- the unsoundness
is fundamentally about the type system's treatment of descriptors
themselves. I created a new "descriptors" category for this one.

I also adjusted the README to use the function name `func` that is
actually required in examples.

---------

Co-authored-by: Jelle Zijlstra <jelle.zijlstra@gmail.com>