Adding new examples in runtime and stdlib (#13)
`dunder_variables.py` shows that the mangled versions of instance
variable names are not properly handed in typing. This allows for the
non-mangled version to be overwritten with incorrect typing and the type
checker does not notice.

`setattr.py` allows for overwriting of a value without proper type
checking. This enables a classes instance variable to be changed to a
different type.