Add example of how modifying __closure__ is unsound (#17)
This requires manually modifying values based on knowledge of the code,
and where captured variables are stored, so it is not very practical or
likely used in any real world scenarios. I 100% will admit that doing
anything like this is A Really Bad Idea™ so feel free to reject if this
is too esoteric.

With that being said, `__closure__` can contain variables that are
referenced in a function, in this case via nonlocal that would have gone
out of scope when `__init__` exits. You can then modify these values
behind the back of the type checker. But if you are modifying
`__closure__`, you are likely doing something you shouldn't be doing.