Add example of calling a method on the wrong class (#33)
I placed it in the `examples/` folder directly because I can't think of
a neat category for this error.

A subclass _technically_ always violates LSP if it overrides a method,
since `self` in the child method has a new restriction on `self` that
the parent didn't have.

This also works when `Fooable` is a protocol.